pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "match-events"
version = "0.1.0"
description = "Broadcast-video football tracking and match-events pipeline"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "MarcosMig" }
]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "opencv-python>=4.9",
  "pyyaml>=6.0",
  "ultralytics>=8.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "black>=24.0"
]
small_objects = [
  "sahi>=0.11.21"
]
soccernet = [
  "SoccerNet>=0.1.62"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
norecursedirs = [
  ".git",
  ".pytest_cache",
  "pytest-cache-files-*",
  "tests/.tmp",
  "runs",
  "outputs",
]
